added some extra functions

+++ ft_strcat.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/11 10:48:45 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/01 12:02:09 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/24 13:42:08 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 char	*ft_strcat(char *str1, const char *str2)
 {
-	int	i;
-	int	len;
+	size_t	i;
+	size_t	len;
 
 	i = 0;
-	len = (int)ft_strlen(str1);
+	len = ft_strlen(str1);
 	while (str2[i] != '\0')
 	{
 		str1[(len + i)] = str2[i];
